+++ libs/journeys/ui/src/components/SearchBar/SearchDropdown/ContinentRefinements/index.ts
@@ -1,0 +1,1 @@
+export { ContinentRefinements } from './ContinentRefinements'
