+++ libs/journeys/ui/src/components/SearchBar/SearchDropdown/SearchbarDropdown.tsx
@@ -2,15 +2,9 @@ import Box from '@mui/material/Box'
 import Popper from '@mui/material/Popper'
 import Stack from '@mui/material/Stack'
 import { useTheme } from '@mui/material/styles'
-import Typography from '@mui/material/Typography'
-import { useTranslation } from 'next-i18next'
-import { ReactElement, ReactNode } from 'react'
-import { useRefinementList } from 'react-instantsearch'
+import { ReactElement } from 'react'
 
-import { useLanguagesContinentsQuery } from '../../../libs/useLanguagesContinentsQuery'
-import { useSortLanguageContinents } from '../../../libs/useSortLanguageContinents'
-
-import { RefinementGroup } from './RefinementGroup'
+import { ContinentRefinements } from './ContinentRefinements'
 
 interface SearchbarDropdownProps {
   open: boolean
@@ -23,49 +17,7 @@ export function SearchbarDropdown({
   id,
   anchorEl
 }: SearchbarDropdownProps): ReactElement {
-  const { t } = useTranslation('apps-watch')
   const theme = useTheme()
-
-  const { data } = useLanguagesContinentsQuery()
-  const languages = useSortLanguageContinents({
-    languages: data?.languages ?? []
-  })
-  const refinements = useRefinementList({
-    attribute: 'languageEnglishName',
-    limit: 1000
-  })
-
-  function ContinentRefinements(): ReactNode {
-    return refinements.items.length > 0 ? (
-      <>
-        {Object.entries(languages).map(([continent, continentLanguages]) => {
-          const items = refinements.items.filter((item) =>
-            continentLanguages.some((language) => language === item.label)
-          )
-
-          return items.length > 0 ? (
-            <RefinementGroup
-              key={continent}
-              title={continent}
-              refinement={{
-                ...refinements,
-                items
-              }}
-            />
-          ) : (
-            <></>
-          )
-        })}
-      </>
-    ) : (
-      <Typography>
-        {t(
-          `Sorry, there are no languages available for this search. Try removing some of your search criteria!`
-        )}
-      </Typography>
-    )
-  }
-
   return (
     <Popper
       id={id}
